% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huttenlocher1991-package.R
\docType{data}
\name{vocab_growth}
\alias{vocab_growth}
\title{Vocabulary growth for 22 toddlers}
\format{
A data frame with 126 rows and 9 variables:
\describe{
  \item{id}{child id. there are 22 unique children}
  \item{mom_speak}{tbd}
  \item{sex}{tbd}
  \item{group}{tbd}
  \item{log_mom}{tbd}
  \item{age}{age in months}
  \item{vocab}{vocabulary, type tbd}
  \item{age_12}{age in months centered at 12 months}
  \item{age_12_sq}{age in months squared centered at 12 months}
}
}
\source{
Downloaded from
  \url{https://ssicentral.com/index.php/products/hlm-general/hlm-examples/}
}
\usage{
vocab_growth
}
\description{
A dataset of children's vocabulary over time.
}
\references{
Huttenlocher, J., Haight, W., Bryk, A., Seltzer, M., & Lyons, T.
  (1991). Early vocabulary growth: relation to language input and gender.
  Developmental psychology, 27(2), 236.
  \url{https://doi.org/10.1037/0012-1649.27.2.236}
}
\keyword{datasets}
